ai_youtube_search_backend/requirements.txt
Flask
Flask-CORS
google-api-python-client
google-generativeai